Add test-plugin
